sequdas_web/requirements.txt
django
psycopg2
django-filters
django-cors-headers
graphene-django
djangorestframework
djangorestframework-jwt
